# Replace encryption algorithm