--- a/secio/benches/bench.rs
+++ b/secio/benches/bench.rs
@@ -2,43 +2,59 @@ use bytes::BytesMut;
 use criterion::{criterion_group, criterion_main, Bencher, Criterion};
 use tentacle_secio::{
     codec::Hmac,
-    stream_cipher::{ctr_init, Cipher},
+    crypto::{cipher::CipherType, new_stream, CryptoMode},
     Digest,
 };
 
-const NULL_IV: [u8; 16] = [0; 16];
-
-fn decode_encode(data: &[u8], cipher: Cipher) {
+fn decode_encode(data: &[u8], cipher: CipherType) {
     let cipher_key = (0..cipher.key_size())
         .map(|_| rand::random::<u8>())
         .collect::<Vec<_>>();
     let hmac_key: [u8; 32] = rand::random();
+    let iv = (0..cipher.iv_size())
+        .map(|_| rand::random::<u8>())
+        .collect::<Vec<_>>();
 
-    let mut encode_data = BytesMut::from(data);
+    let mut encode_cipher = new_stream(cipher, &cipher_key, &iv, CryptoMode::Encrypt);
+    let mut decode_cipher = new_stream(cipher, &cipher_key, &iv, CryptoMode::Decrypt);
+    let (mut decode_hmac, mut encode_hmac) = match cipher {
+        CipherType::ChaCha20Poly1305 | CipherType::Aes128Gcm | CipherType::Aes256Gcm => (None, None),
+        _ => {
+            let encode_hmac = Hmac::from_key(Digest::Sha256, &hmac_key);
+            let decode_hmac = encode_hmac.clone();
+            (Some(decode_hmac), Some(encode_hmac))
+        }
+    };
 
-    let mut encode_cipher = ctr_init(cipher, &cipher_key, &NULL_IV);
-    let mut encode_hmac = Hmac::from_key(Digest::Sha256, &hmac_key);
-    let mut decode_cipher = ctr_init(cipher, &cipher_key, &NULL_IV);
-    let mut decode_hmac = encode_hmac.clone();
+    let mut encode_data = BytesMut::new();
+    encode_cipher.update(&data[..], &mut encode_data).unwrap();
+    if encode_hmac.is_some() {
+        let signature = encode_hmac.as_mut().unwrap().sign(&encode_data[..]);
+        encode_data.extend_from_slice(signature.as_ref());
+    }
 
-    encode_cipher.encrypt(&mut encode_data[..]);
-    let signature = encode_hmac.sign(&encode_data[..]);
-    encode_data.extend_from_slice(signature.as_ref());
+    if decode_hmac.is_some() {
+        let content_length = encode_data.len() - decode_hmac.as_mut().unwrap().num_bytes();
 
-    let content_length = encode_data.len() - decode_hmac.num_bytes();
+        let (crypted_data, expected_hash) = encode_data.split_at(content_length);
 
-    let (crypted_data, expected_hash) = encode_data.split_at(content_length);
+        assert!(decode_hmac
+            .as_mut()
+            .unwrap()
+            .verify(crypted_data, expected_hash));
 
-    assert!(decode_hmac.verify(crypted_data, expected_hash));
+        encode_data.truncate(content_length);
+    }
 
-    let mut decode_data = encode_data.to_vec();
-    decode_data.truncate(content_length);
-    decode_cipher.decrypt(&mut decode_data);
+    let mut decode_data = BytesMut::new();
+    decode_cipher
+        .update(&encode_data, &mut decode_data)
+        .unwrap();
 
     assert_eq!(&decode_data[..], &data[..]);
 }
 
-fn bench_test(bench: &mut Bencher, cipher: Cipher, data: &[u8]) {
+fn bench_test(bench: &mut Bencher, cipher: CipherType, data: &[u8]) {
     bench.iter(|| {
         decode_encode(data, cipher);
     })
@@ -48,31 +64,47 @@ fn criterion_benchmark(bench: &mut Criterion) {
     let data = (0..1024 * 256)
         .map(|_| rand::random::<u8>())
         .collect::<Vec<_>>();
-    bench.bench_function("1kb_aes128", {
+    bench.bench_function("1kb_aes128ctr", {
         let data = data.clone();
-        move |b| bench_test(b, Cipher::Aes128, &data)
+        move |b| bench_test(b, CipherType::Aes128Ctr, &data)
     });
-    bench.bench_function("1kb_aes256", {
+    bench.bench_function("1kb_aes256ctr", {
         let data = data.clone();
-        move |b| bench_test(b, Cipher::Aes256, &data)
+        move |b| bench_test(b, CipherType::Aes256Ctr, &data)
     });
-    bench.bench_function("1kb_twofish", move |b| {
-        bench_test(b, Cipher::TwofishCtr, &data)
+    bench.bench_function("1kb_aes128gcm", {
+        let data = data.clone();
+        move |b| bench_test(b, CipherType::Aes128Gcm, &data)
+    });
+    bench.bench_function("1kb_aes256gcm", {
+        let data = data.clone();
+        move |b| bench_test(b, CipherType::Aes256Gcm, &data)
+    });
+    bench.bench_function("1kb_chacha20poly1305", move |b| {
+        bench_test(b, CipherType::ChaCha20Poly1305, &data)
     });
 
     let data = (0..1024 * 1024)
         .map(|_| rand::random::<u8>())
         .collect::<Vec<_>>();
-    bench.bench_function("1mb_aes128", {
+    bench.bench_function("1mb_aes128ctr", {
         let data = data.clone();
-        move |b| bench_test(b, Cipher::Aes128, &data)
+        move |b| bench_test(b, CipherType::Aes128Ctr, &data)
     });
-    bench.bench_function("1mb_aes256", {
+    bench.bench_function("1mb_aes256ctr", {
         let data = data.clone();
-        move |b| bench_test(b, Cipher::Aes256, &data)
+        move |b| bench_test(b, CipherType::Aes256Ctr, &data)
     });
-    bench.bench_function("1mb_twofish", move |b| {
-        bench_test(b, Cipher::TwofishCtr, &data)
+    bench.bench_function("1mb_aes128gcm", {
+        let data = data.clone();
+        move |b| bench_test(b, CipherType::Aes128Gcm, &data)
+    });
+    bench.bench_function("1mb_aes256gcm", {
+        let data = data.clone();
+        move |b| bench_test(b, CipherType::Aes256Gcm, &data)
+    });
+    bench.bench_function("1mb_chacha20poly1305", move |b| {
+        bench_test(b, CipherType::ChaCha20Poly1305, &data)
     });
 }