… secio/Cargo.toml
… secio/src/crypto/cipher.rs
… secio/src/crypto/mod.rs
… secio/src/crypto/openssl_impl.rs
… secio/src/crypto/ring_impl.rs
… secio/src/error.rs
… secio/src/handshake/handshake_context.rs
… secio/src/handshake/mod.rs
… secio/src/handshake/procedure.rs
… secio/src/lib.rs
… secio/src/peer_id.rs
… secio/src/stream_cipher.rs
… secio/src/support.rs
… src/utils.rs
… tests/test_peer_id.rs