@driftluo 2019-10-16
Add a contrasting implementation of openssl and ring, and use the openssl implementation by default with native support

After testing, openssl implementation will be 5-20% faster than ring on my machine.